bug: config parsing format fixed
Config parser has been updated to work with the updated config file
syntax.

I decided to remove the support for the version in the project's
pdf, since it's useless. All it will do is make the regular expression
harder to debug and understand, also the YAGNI principle ;)

Tests and old config files have been updated.

closes #8